requirements.txt
notion-client>=2.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
